pyproject.toml
[project]
name = "smartstay-intelligence"
version = "0.1.0"
description = "AI-powered revenue optimization platform for hotel management"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
keywords = ["hotel", "revenue-management", "machine-learning", "fastapi", "xgboost"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web Framework
    "fastapi[standard]>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    # Database
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    # Cache
    "redis>=5.0.1",
    # Machine Learning
    "xgboost>=2.0.3",
    "prophet>=1.1.5",
    "scikit-learn>=1.4.0",
    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    # HTTP Requests & Web Scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.26.0",
    # Configuration
    "python-dotenv>=1.0.0",
    "bcrypt==3.2.2",
    "openpyxl>=3.1.5",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "apify-client>=2.5.0",
    "vadersentiment>=3.3.2",
    "google-generativeai>=0.8.6",
    "mlflow>=3.9.0",
    "lightgbm>=4.6.0",
]

[project.optional-dependencies]
# Développement
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.4",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
]

# MLOps
mlops = [
    "mlflow>=2.10.0",
    "evidently>=0.4.13",
]

# Airflow (pipelines)
airflow = [
    "apache-airflow>=2.8.0",
]

# Production monitoring
monitoring = [
    "prometheus-client>=0.19.0",
    "sentry-sdk[fastapi]>=1.40.0",
]

# All optional dependencies
all = [
    "smartstay-intelligence[dev,mlops,airflow,monitoring]",
]

[project.urls]
Homepage = "https://github.com/yourusername/smartstay-intelligence"
Documentation = "https://github.com/yourusername/smartstay-intelligence#readme"
Repository = "https://github.com/yourusername/smartstay-intelligence"
Issues = "https://github.com/yourusername/smartstay-intelligence/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--cov=backend/app",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--verbose",
    "--strict-markers",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
]
max-complexity = 10

[tool.coverage.run]
source = ["backend/app"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
